moved most of nav state to a context so it can be shared with a new Option component
