+++ src/components.rs
@@ -2,3 +2,4 @@ pub mod card;
 pub mod contact;
 pub mod hero;
 pub mod nav;
+pub mod option;
