+++ src/components/nav.rs
@@ -1,86 +1,16 @@
+// use crate::components::option::Option;
 use sycamore::prelude::*;
 use sycamore_router::navigate;
 
-#[derive(Clone, Copy, PartialEq, Eq)]
-struct TabStateData {
-    selected_anchor_classes: &'static str,
-    unselected_anchor_classes: &'static str,
-    selected_span_classes: &'static str,
-    unselected_span_classes: &'static str,
-}
-
-impl TabStateData {
-    fn new() -> Self {
-        Self {
-            selected_anchor_classes: "bg-gray-100 text-gray-900 group relative min-w-0 flex-1 overflow-hidden py-4 px-4 text-center text-sm font-medium hover:bg-sky-100 focus:z-10",
-            unselected_anchor_classes: "bg-gray-100 text-gray-400 hover:text-gray-700 group relative min-w-0 flex-1 overflow-hidden py-4 px-4 text-center text-sm font-medium hover:bg-sky-100 focus:z-10",
-            selected_span_classes: "bg-pink-500 absolute inset-x-0 bottom-0 h-0.5",
-            unselected_span_classes: "bg-transparent absolute inset-x-0 bottom-0 h-0.5",
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-struct CurrentTabState {
-    project_anchor_classes: &'static str,
-    project_span_classes: &'static str,
-    about_anchor_classes: &'static str,
-    about_span_classes: &'static str,
-    contact_anchor_classes: &'static str,
-    contact_span_classes: &'static str,
-}
-
-impl CurrentTabState {
-    fn new() -> Self {
-        Self {
-            project_anchor_classes: "",
-            project_span_classes: "",
-            about_anchor_classes: "",
-            about_span_classes: "",
-            contact_anchor_classes: "",
-            contact_span_classes: "",
-        }
-    }
-
-    fn select_project(&self, tab_state_data: &TabStateData) -> Self {
-        Self {
-            project_anchor_classes: tab_state_data.selected_anchor_classes,
-            project_span_classes: tab_state_data.selected_span_classes,
-            about_anchor_classes: tab_state_data.unselected_anchor_classes,
-            about_span_classes: tab_state_data.unselected_span_classes,
-            contact_anchor_classes: tab_state_data.unselected_anchor_classes,
-            contact_span_classes: tab_state_data.unselected_span_classes,
-        }
-    }
-    fn select_about(&self, tab_state_data: &TabStateData) -> Self {
-        Self {
-            project_anchor_classes: tab_state_data.unselected_anchor_classes,
-            project_span_classes: tab_state_data.unselected_span_classes,
-            about_anchor_classes: tab_state_data.selected_anchor_classes,
-            about_span_classes: tab_state_data.selected_span_classes,
-            contact_anchor_classes: tab_state_data.unselected_anchor_classes,
-            contact_span_classes: tab_state_data.unselected_span_classes,
-        }
-    }
-    fn select_contact(&self, tab_state_data: &TabStateData) -> Self {
-        Self {
-            project_anchor_classes: tab_state_data.unselected_anchor_classes,
-            project_span_classes: tab_state_data.unselected_span_classes,
-            about_anchor_classes: tab_state_data.unselected_anchor_classes,
-            about_span_classes: tab_state_data.unselected_span_classes,
-            contact_anchor_classes: tab_state_data.selected_anchor_classes,
-            contact_span_classes: tab_state_data.selected_span_classes,
-        }
-    }
-}
+use crate::CurrentTabState;
+use crate::TabStateData;
 
 #[component]
 pub fn Nav<G: Html>(cx: Scope) -> View<G> {
     let select_state = create_signal(cx, "/");
-    let tab_state_data = TabStateData::new();
-    // the default route is / with the projects tab selected
-    let current_tab_state = CurrentTabState::new().select_project(&tab_state_data);
-    let tab_state = create_signal(cx, current_tab_state);
+    let tab_state_data: &TabStateData = use_context(cx);
+    let current_tab_state: &CurrentTabState = use_context(cx);
+    let tab_state = create_signal(cx, *current_tab_state);
 
     let select_option = |path| select_state.set(path);
     let select_tab = move |path| match path {
@@ -101,6 +31,7 @@ pub fn Nav<G: Html>(cx: Scope) -> View<G> {
                     option(on:click=move |_| {select_option("/"); select_tab("/")}) {"Projects"}
                     option(on:click=move |_| {select_option("/about"); select_tab("/about")}) {"About"}
                     option(on:click=move |_| {select_option("/contact"); select_tab("/contact")}) {"Contact"}
+                    // Option(selected=create_signal(cx, true), label="Bar", path="/about")
                 }
             }
             // tab elements when viewed on medium or larger devices
